langchain-projects/qna_chatbot_openai/requirements.txt
langchain_openai
pandas 
numpy
python-dotenv
langchain_core
langchain_community
streamlit
openai